feat(lifecycle): refactor CLI daemon lifecycle commands using structured subcommands
- Removed legacy lifecycle handler code in favor of new `CliCommand` and `DaemonAction` enums
- Moved lifecycle handling logic into `CliRunner` abstraction for cleaner run and test paths
- Split lifecycle management into focused submodules: types, error, utils, controller
- Extracted lifecycle error types and utilities for better maintainability
- Replaced previous domain/operation parsing with clap subcommands for daemon lifecycle actions (start, stop, status)
- Updated tests and feature scenarios to exercise new CLI lifecycle command structure
- Removed `parse_lifecycle_invocation` function and legacy lifecycle CLI args

This improves CLI ergonomics and internal code structure for daemon lifecycle operations.
